==> S1_Pretraitements.ipynb <==
# On importe la librairie pandas et le dataset source et le dataset TEST
import pandas as pd
from matplotlib import pyplot
# On importe les fonctions depuis my_functions.ipynb
import ipynb.fs.full.my_functions as myfn

# On omporte les csv
df = pd.read_csv('sources/kaggle-give-me-credit-train.csv')
df_test = pd.read_csv('sources/kaggle-give-me-credit-test.csv')

df = myfn.pretraitements(df)
df_test = myfn.pretraitements(df_test)

df_test = myfn.reorganiser_colonnes(df, df_test)

# On exporte en csv
df.to_csv('storage/source_dataset_train_full.csv', index=False)
df_test.to_csv('storage/source_dataset_test.csv', index=False)
