==> S2_BigML_One_Prediction.ipynb <==
# Data to predict
predicton_data = {
    "RevolvingUtilizationOfUnsecuredLines": 0.01703559,
    "NumberOfDependents": 1,
    "DebtRatio": 0,
    "age": 42,
    "NumberOfOpenCreditLinesAndLoans": 6,
    "NumberRealEstateLoansOrLines": 1,
    "NumberOfTime30-59DaysPastDueNotWorse": 1,
    "NumberOfTime60-89DaysPastDueNotWorse": 0,
    "NumberOfTimes90DaysLate": 0,
    "MonthlyIncome" : 100000
}

# On importe la librairie bigml et pandas
from bigml.api import BigML
import pandas as pd
# Importe les fonctions depuis my_functions.ipynb
import ipynb.fs.full.my_functions as myfn

# On se connecte à bigml et on récupère un ensemble
api = BigML(project='project/5d94a32e42129f2e16000232')
ensemble = api.get_ensemble('ensemble/5dc2f3f07811dd7f2d012312')

# On récupère le dataset utilisé pour entrainer le modèle
dataset = api.get_dataset(ensemble['object']['dataset'])

# On récupère l'ensemble des colonnes du dataset
dataset_columns = []
for row in dataset['object']['fields']:
    dataset_columns.append(dataset['object']['fields'][row]['name'])

# On crée un dataframe pandas et on injecte les colonnes
df = pd.DataFrame(columns=dataset_columns)
# puis les données
df = df.append(predicton_data,ignore_index=True)

# Applique les prétraitements sur le dataframe
df = myfn.pretraitements(df)

# Transforme le dataframe en dictionnaire
data = df.iloc[0].to_dict()

myfn.bprint(data)

# Requête BigMl pour obtenir la prediction
prediction = api.create_prediction(ensemble, data, {"explain" : True, "name": "TEST 100000$ monthly income"})
api.ok(prediction)

if prediction['object']['probabilities'][0][1] > 0.5:
    print(f"Prediction is {prediction['object']['probabilities'][0][0]} with a probability of {prediction['object']['probabilities'][0][1]*100}%")
else:
    print(f"Prediction is {prediction['object']['probabilities'][1][0]} with a probability of {prediction['object']['probabilities'][1][1]*100}%")

api.pprint(prediction['object']['explanation'])

liste = []
for row in prediction['object']['explanation']:
    row['field'] = ensemble['object']['ensemble']['fields'][row['field']]['name']
    liste.append(row)
    
print(liste)

for row in prediction['object']['fields']:
    print(prediction['object']['fields'][row]['name'])
