==> S3_Kaggle.ipynb <==
# On importe la librairie pandas et le dataset source
import pandas as pd
# On importe la librairie kaggle
import kaggle

# On prépare le fichier pour kaggle
df = pd.read_csv('storage/batch_prediction.csv')
# On supprime les colonnes inutiles pour kaggle
df.drop(columns=['SeriousDlqin2yrs', '0 probability'], inplace=True)
# On renomme les colonnes au format kaggle
df.rename(columns={'1 probability': 'Probability'}, inplace=True)
# On exporte en csv
df.to_csv('storage/kaggle_submission_file.csv', index=False)

# On soumet à Kaggle
submission_file = "storage/kaggle_submission_file.csv"
df.to_csv(submission_file, index=False)
kaggle.api.competition_submit(submission_file, "Ensemble", "GiveMeSomeCredit")
