==> S5_AUC.ipynb <==
import pandas as pd
df = pd.read_csv('storage/batch_prediction_all_fields.csv', index_col=False, header=0)

nb_inf = 0
nb_interation = 0
serie = df.transpose() # here we convert the DataFrame into a Serie
for ligne in serie: # Pour chaque ligne du tableau serie
    if serie[ligne][11] == 1: # si la valeur de la colonne SeriousDlqin2yrs = 1 alors ...
        proba_positif = serie[ligne][15] # On stocke la valeur de 1 probability à la variable proba_positif
        for ligne_neg in serie: # On boucle sur toutes les lignes da la serie 
            if serie[ligne_neg][11] == 0: # Si la valeur de SeriousDlqin2yrs = 0 alors ...
                if proba_positif > serie[ligne_neg][15]: # Si la valeur de proba_positif est inférieure à la valeur de 1 probability de la ligne (dont SeriousDlqin2yrs est égal à 0 proba_negtive) alors ...
                    nb_inf += 1 # On incrémente la valeur de nb_inf de 1
                nb_interation += 1 # On incrémente le nombre de boucle total (dénominateur pour le calcul de l'AUC)
                    
print(nb_inf)
print(nb_interation)
print(f"{round(nb_inf/nb_interation*100,2)}%")

import pandas as pd
df = pd.read_csv('storage/batch_prediction_all_fields.csv', index_col=False, header=0)
pos = df.loc[df['SeriousDlqin2yrs'] == 1] # Tout les Positif 
neg = df.loc[df['SeriousDlqin2yrs'] == 0] # Tout les negatifs 

x = 0
y = 0
# pour chaque 1 proba ( P ) dans toutes les 1 proba (P)
for threshold_pos in pos['1 probability']:
    # pour chaque 1 Proba ( N ) dans toutes les 1 proba ( N )
    for threshold_neg in neg['1 probability']:
        if threshold_pos > threshold_neg:
            x += 1 
        y += 1
    
print(x)
print(y)
print(f"{round((x/y)*100,2)}%")

import pandas as pd
df = pd.read_csv('storage/batch_prediction_all_fields.csv', index_col=False, header=0)
pos = df.loc[df['SeriousDlqin2yrs']  == 0]
print(pos)
